+++ src/api/services/users.service.ts
@@ -1,9 +1,5 @@
-import Sequelize from 'sequelize'
-
 import db from '../../database/models'
 import { UserInstance } from '../../database/models/users'
-
-const { Op } = Sequelize
 
 /**
  * @export
